refactor(middleware-io): moved http and event validators to separate files
