+++ packages/middleware-io/src/index.ts
@@ -1,6 +1,8 @@
 import * as io from 'io-ts';
 
 // package public API
-export { Schema, ValidatorOptions, validator$, httpValidator$, eventValidator$ } from './io.middleware';
+export { Schema, ValidatorOptions, validator$ } from './io.middleware';
+export { httpValidator$ } from './io.http.middleware';
+export { eventValidator$ } from './io.event.middleware';
 export { defaultReporter } from './io.reporter';
 export { io };
